+++ .github/workflows/release.yml
@@ -13,7 +13,7 @@ jobs:
 
     environment:
       name: pypi
-      url: https://pypi.org/p/sole
+      url: https://pypi.org/p/sour
 
     permissions:
       id-token: write
