+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "sole"
+name = "sour"
 version = "0.1.0"
 description = "A tool for dynamically generating markdown files using references to commands."
 readme = "README.md"
@@ -15,8 +15,8 @@ classifiers = [
     "Topic :: Software Development :: Documentation",
     "Topic :: Text Processing :: Markup :: Markdown",
 ]
-keywords = ["documentation", "markdown", "generator", "sole", "devops"]
-urls = { Repository = "https://github.com/Solenya-AIaaS/sole", Issues = "https://github.com/Solenya-AIaaS/sole/issues" }
+keywords = ["documentation", "markdown", "generator", "sour", "devops"]
+urls = { Repository = "https://github.com/Solenya-AIaaS/sour", Issues = "https://github.com/Solenya-AIaaS/sour/issues" }
 
 dependencies = [
     "typer>=0.9.0",
@@ -33,14 +33,14 @@ dev = [
 ]
 
 [project.scripts]
-sole = "sole.main:app"
+sour = "sour.main:app"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/sole"]
+packages = ["src/sour"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
