+++ tests/test_version.py
@@ -1,4 +1,4 @@
-from sole import __version__
+from sour import __version__
 
 
 def test_version():
